Create Watson Deep Learning For Find Missing Persons Using IBM Cloud
In this scenario, missing case entries are updated with their photocopies in the police station. By using CCTV cameras we can compare each person with the available database and find these people. If the missing person is found in the CCTV Video stream then location which is tagged to the CCTV is sent as an SMS to the police station.